requirement.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
appdirs==1.4.4
asttokens==2.4.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==7.1.0
importlib-resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
jupyter-client==8.6.1
jupyter-core==5.7.2
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmcv==2.1.0
mmengine==0.10.4
mmsegmentation==1.2.2
model-index==0.1.11
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.1
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.38
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.1
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.4.28
requests==2.28.1
rich==13.4.2
safetensors==0.4.3
scipy==1.10.1
sentry-sdk==2.0.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
termcolor==2.4.0
thop==0.1.1.post2209072238
timm==0.9.16
tomli==2.0.1
torch==2.1.2+cu118
torcheval==0.0.7
tornado==6.4
tqdm==4.65.2
traitlets==5.14.3
typing-extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.6
wcwidth==0.2.13
yapf==0.40.2
zipp==3.18.1
